Move functions for generating Gethdata to a new crate mock.
