+++ Cargo.toml
@@ -7,7 +7,8 @@ members = [
     "integration-tests",
     "circuit-benchmarks",
     "eth-types",
-    "external-tracer"
+    "external-tracer",
+    "mock"
 ]
 
 [patch.crates-io]
