+++ bus-mapping/Cargo.toml
@@ -16,7 +16,7 @@ serde = {version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.66"
 
 [dev-dependencies]
-external-tracer = { path = "../external-tracer" }
+mock = { path = "../mock" }
 pretty_assertions = "1.0.0"
 tokio = { version = "1.13", features = ["macros"] }
 url = "2.2.2"
