+++ bus-mapping/src/evm/opcodes/coinbase.rs
@@ -25,13 +25,9 @@ impl Opcode for Coinbase {
 #[cfg(test)]
 mod coinbase_tests {
     use super::*;
-    use crate::{
-        bytecode,
-        circuit_input_builder::{ExecStep, TransactionContext},
-        mock,
-    };
+    use crate::circuit_input_builder::{ExecStep, TransactionContext};
     use eth_types::evm_types::StackAddress;
-    use eth_types::ToWord;
+    use eth_types::{bytecode, ToWord};
     use pretty_assertions::assert_eq;
 
     #[test]
@@ -43,7 +39,9 @@ mod coinbase_tests {
         };
 
         // Get the execution steps from the external tracer
-        let block = mock::BlockData::new_single_tx_trace_code_at_start(&code).unwrap();
+        let block = crate::mock::BlockData::new_from_geth_data(
+            mock::new_single_tx_trace_code_at_start(&code).unwrap(),
+        );
 
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
