+++ bus-mapping/src/evm/opcodes/msize.rs
@@ -29,13 +29,9 @@ impl Opcode for Msize {
 #[cfg(test)]
 mod msize_tests {
     use super::*;
-    use crate::{
-        bytecode,
-        circuit_input_builder::{ExecStep, TransactionContext},
-        mock,
-    };
+    use crate::circuit_input_builder::{ExecStep, TransactionContext};
     use eth_types::evm_types::StackAddress;
-    use eth_types::Word;
+    use eth_types::{bytecode, Word};
 
     #[test]
     fn msize_opcode_impl() -> Result<(), Error> {
@@ -47,7 +43,9 @@ mod msize_tests {
             STOP
         };
 
-        let block = mock::BlockData::new_single_tx_trace_code_at_start(&code).unwrap();
+        let block = crate::mock::BlockData::new_from_geth_data(
+            mock::new_single_tx_trace_code_at_start(&code).unwrap(),
+        );
 
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
