+++ bus-mapping/src/evm/opcodes/stackonlyop.rs
@@ -38,13 +38,9 @@ impl<const N: usize> Opcode for StackOnlyOpcode<N> {
 #[cfg(test)]
 mod stackonlyop_tests {
     use super::*;
-    use crate::{
-        bytecode,
-        circuit_input_builder::{ExecStep, TransactionContext},
-        mock,
-    };
+    use crate::circuit_input_builder::{ExecStep, TransactionContext};
     use eth_types::evm_types::StackAddress;
-    use eth_types::{word, Word};
+    use eth_types::{bytecode, word, Word};
     use pretty_assertions::assert_eq;
 
     #[test]
@@ -57,7 +53,9 @@ mod stackonlyop_tests {
         };
 
         // Get the execution steps from the external tracer
-        let block = mock::BlockData::new_single_tx_trace_code_at_start(&code).unwrap();
+        let block = crate::mock::BlockData::new_from_geth_data(
+            mock::new_single_tx_trace_code_at_start(&code).unwrap(),
+        );
 
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
@@ -115,7 +113,9 @@ mod stackonlyop_tests {
         };
 
         // Get the execution steps from the external tracer
-        let block = mock::BlockData::new_single_tx_trace_code_at_start(&code).unwrap();
+        let block = crate::mock::BlockData::new_from_geth_data(
+            mock::new_single_tx_trace_code_at_start(&code).unwrap(),
+        );
 
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
@@ -175,7 +175,9 @@ mod stackonlyop_tests {
         };
 
         // Get the execution steps from the external tracer
-        let block = mock::BlockData::new_single_tx_trace_code_at_start(&code).unwrap();
+        let block = crate::mock::BlockData::new_from_geth_data(
+            mock::new_single_tx_trace_code_at_start(&code).unwrap(),
+        );
 
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
