+++ zkevm-circuits/Cargo.toml
@@ -24,7 +24,7 @@ rand = "0.8"
 
 [dev-dependencies]
 criterion = "0.3"
-external-tracer = { path = "../external-tracer" }
+mock = { path = "../mock" }
 
 [[bench]]
 name = "binary_value"
