+++ zkevm-circuits/src/evm_circuit/execution/add.rs
@@ -104,9 +104,8 @@ impl<F: FieldExt> ExecutionGadget<F> for AddGadget<F> {
 mod test {
     use crate::evm_circuit::test::rand_word;
     use crate::test_util::run_test_circuits;
-    use bus_mapping::bytecode;
     use eth_types::evm_types::OpcodeId;
-    use eth_types::Word;
+    use eth_types::{bytecode, Word};
 
     fn test_ok(opcode: OpcodeId, a: Word, b: Word) {
         let bytecode = bytecode! {
