+++ zkevm-circuits/src/evm_circuit/execution/jump.rs
@@ -89,7 +89,7 @@ impl<F: FieldExt> ExecutionGadget<F> for JumpGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::{evm_circuit::test::rand_range, test_util::run_test_circuits};
-    use bus_mapping::bytecode;
+    use eth_types::bytecode;
 
     fn test_ok(destination: usize) {
         assert!((34..(1 << 24) - 1).contains(&destination));
