+++ zkevm-circuits/src/evm_circuit/execution/msize.rs
@@ -77,8 +77,7 @@ impl<F: FieldExt> ExecutionGadget<F> for MsizeGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::test_util::run_test_circuits;
-    use bus_mapping::bytecode;
-    use eth_types::Word;
+    use eth_types::{bytecode, Word};
 
     #[test]
     fn msize_gadget() {
