+++ zkevm-circuits/src/evm_circuit/execution/pc.rs
@@ -75,7 +75,7 @@ impl<F: FieldExt> ExecutionGadget<F> for PcGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::test_util::run_test_circuits;
-    use bus_mapping::bytecode;
+    use eth_types::bytecode;
 
     fn test_ok() {
         let bytecode = bytecode! {
